utility: use static inline for some functions
This changes many functions from internal.c into static inline since they are just wrappers with extra asserts.
This change should improve the compiler's ability to perform optimizations, since it now always has insight into the functions.

The static inline functions are only declared in internal headers.
To make them available publically, we also re-export them via elektra-utility.so.
