+++ src/libs/base/internal.c
@@ -44,74 +44,6 @@
  *
  */
 
-/**@brief Compare Strings.
- *
- * @param s1 The first string to be compared
- * @param s2 The second string to be compared
- *
- * @ingroup internal
- * @return a negative number if s1 is less than s2
- * @retval 0 if s1 matches s2
- * @return a positive number if s1 is greater than s2
- **/
-int elektraStrCmp (const char * s1, const char * s2)
-{
-	ELEKTRA_ASSERT (s1 != NULL && s2 != NULL, "Got null pointer s1: %p s2: %p", (void *) s1, (void *) s2);
-
-	return strcmp (s1, s2);
-}
-
-/**@brief Compare Strings up to a maximum length.
- *
- * @param s1 The first string to be compared
- * @param s2 The second string to be compared
- * @param n The maximum length to be compared
- *
- * @ingroup internal
- * @return a negative number if s1 is less than s2
- * @retval 0 if s1 matches s2
- * @return a positive number if s1 is greater than s2
- **/
-int elektraStrNCmp (const char * s1, const char * s2, size_t n)
-{
-	ELEKTRA_ASSERT (s1 != NULL && s2 != NULL, "Got null pointer s1: %p s2: %p", (void *) s1, (void *) s2);
-
-	return strncmp (s1, s2, n);
-}
-
-/**@brief Compare Strings ignoring case.
- *
- * @param s1 The first string to be compared
- * @param s2 The second string to be compared
- *
- * @ingroup internal
- * @return a negative number if s1 is less than s2
- * @retval 0 if s1 matches s2
- * @return a positive number if s1 is greater than s2
- **/
-int elektraStrCaseCmp (const char * s1, const char * s2)
-{
-	ELEKTRA_ASSERT (s1 != NULL && s2 != NULL, "Got null pointer s1: %p s2: %p", (void *) s1, (void *) s2);
-	return strcasecmp (s1, s2);
-}
-
-/**@brief Compare Strings ignoring case up to a maximum length.
- *
- * @param s1 The first string to be compared
- * @param s2 The second string to be compared
- * @param n The maximum length to be compared
- *
- * @ingroup internal
- * @return a negative number if s1 is less than s2
- * @retval 0 if s1 matches s2
- * @return a positive number if s1 is greater than s2
- **/
-int elektraStrNCaseCmp (const char * s1, const char * s2, size_t n)
-{
-	ELEKTRA_ASSERT (s1 != NULL && s2 != NULL, "Got null pointer s1: %p s2: %p", (void *) s1, (void *) s2);
-	return strncasecmp (s1, s2, n);
-}
-
 /**
  * @brief Compare two memory regions but make cmp chars uppercase before
  * comparison.
@@ -140,188 +72,6 @@ int elektraMemCaseCmp (const char * s1, const char * s2, size_t size)
 	}
 	return 0;
 }
-
-/**Reallocate Storage in a save way.
- *
- *@code
-if (elektraRealloc ((void **) & buffer, new_length) < 0) {
-	// here comes the failure handler
-	// you can still use the old buffer
-#if DEBUG
-	fprintf (stderr, "Reallocation error\n");
-#endif
-	elektraFree (buffer);
-	buffer = 0;
-	// return with error
-}
- *@endcode
- *
- * @param buffer is a pointer to a elektraMalloc
- * @param size is the new size for the memory
- * @retval -1 on failure
- * @retval 0 on success
- * @ingroup internal
- */
-int elektraRealloc (void ** buffer, size_t size)
-{
-	ELEKTRA_ASSERT (size, "Size to allocate is zero (implementation defined behavior)");
-	void * ptr;
-	void * svr = *buffer;
-
-	ptr = realloc (*buffer, size);
-	ELEKTRA_ASSERT (ptr, "Memory (re)allocation failed with size %zu", size);
-	if (ptr == NULL)
-	{
-		*buffer = svr; /* restore old buffer*/
-		return -1;
-	}
-	else
-	{
-		*buffer = ptr;
-		return 0;
-	}
-}
-
-/**
- * Allocate memory for Elektra.
- *
- * @code
-if ((buffer = elektraMalloc (length)) == 0) {
-	// here comes the failure handler
-	// no allocation happened here, so don't use buffer
-#if DEBUG
-	fprintf (stderr, "Allocation error");
-#endif
-	// return with error
-}
- * @endcode
- *
- * @param size the requested size
- *
- * This function is compatible to ANSI-C malloc
- * @see elektraFree
- * @see elektraCalloc
- */
-void * elektraMalloc (size_t size)
-{
-	ELEKTRA_ASSERT (size, "Size to allocate is zero (implementation defined behavior)");
-	void * ret = malloc (size);
-	ELEKTRA_ASSERT (ret, "Memory allocation failed with size %zu", size);
-	return ret;
-}
-
-/**Allocate memory for Elektra.
- *
- * Memory will be set to 0.
- *
- * @param size the requested size
- * @see elektraMalloc
- */
-void * elektraCalloc (size_t size)
-{
-	ELEKTRA_ASSERT (size, "Size to allocate is zero (implementation defined behavior)");
-	void * ret = calloc (1, size);
-	ELEKTRA_ASSERT (ret, "Memory allocation failed with size %zu", size);
-	return ret;
-}
-
-/**Free memory of Elektra or its backends.
- *
- * @param ptr the pointer to free
- *
- * If ptr is NULL, no operation is performed.
- *
- * @ingroup internal
- * @see elektraMalloc
- */
-void elektraFree (void * ptr)
-{
-	free (ptr);
-}
-
-
-/**Copy string into new allocated memory.
- *
- * You need to free the memory yourself.
- *
- * @note that size is determined at runtime.
- *       So if you have a size information, don't use
- *       that function.
- *
- * @param s the null-terminated string to duplicate
- *
- * @ingroup internal
- * @return 0 if out of memory, a pointer otherwise
- * @pre s must be a c-string.
- * @see elektraFree
- * @see elektraStrLen
- * @see elektraMemDup
- */
-char * elektraStrDup (const char * s)
-{
-	void * tmp = 0;
-	size_t l = 0;
-	ELEKTRA_ASSERT (s, "Tried to duplicate null pointer");
-
-	l = elektraStrLen (s);
-	ELEKTRA_ASSERT (l, "Size of string to duplicate is zero");
-	tmp = elektraMalloc (l);
-	if (tmp) memcpy (tmp, s, l);
-
-	return tmp;
-}
-
-/**Copy buffer into new allocated memory.
- *
- * You need to free the memory yourself.
- *
- * This function also works with \\0 characters
- * in the buffer. The length is taken as given,
- * it must be correct.
- *
- * @return 0 if out of memory, a pointer otherwise
- * @param s must be an allocated buffer
- * @param n the number of bytes to copy from s
- * @ingroup internal
- */
-void * elektraMemDup (const void * s, size_t n)
-{
-	void * tmp = 0;
-	ELEKTRA_ASSERT (n, "Size of memory to duplicate is zero");
-
-	tmp = elektraMalloc (n);
-	if (tmp) memcpy (tmp, s, n);
-
-	return tmp;
-}
-
-
-/**
- * Calculates the length in bytes of a string.
- *
- * This function differs from strlen() because it is Unicode and multibyte
- * chars safe. While strlen() counts characters and ignores the final NULL,
- * elektraStrLen() count bytes including the ending NULL.
- *
- * It must not be used to search for / in the name, because it does not
- * consider escaping. Instead use the unescaped name.
- *
- * @see keyUnescapedName()
- *
- * @ingroup internal
- * @param s the string to get the length from
- * @return number of bytes used by the string, including the final NULL.
- * @ingroup internal
- */
-size_t elektraStrLen (const char * s)
-{
-	ELEKTRA_ASSERT (s, "Got null pointer");
-
-	char * found = strchr (s, 0);
-	if (found) return found - s + 1;
-	return 0;
-}
-
 
 /**
  * @brief Does string formatting in fresh allocated memory
