+++ src/libs/utility/alloc.c
@@ -1,0 +1,5 @@
+#include <elektra/utility/alloc.h>
+
+// the public header declares the functions as normal external functions
+// we now include the implementations to actually define them
+#include <internal/utility/alloc_impl.h>
